Finalize IDF sync guardrail with schema update and propagation fixes
- Modified database.sql: changed idf_positions.position_no to smallint to support up to 250 positions and temporary staging offsets.
- Enhanced link_create.php and port_update.php: ensured full metadata propagation (status, color, labels, notes) to switch_ports and peer idf_ports.
- Removed COALESCE patterns in sync updates to allow clearing fields.
- Fixed parameter binding types and order in port_update.php.
- Added error handling to move/reorder batch updates.
- Updated AGENTS.md with high-density support documentation.
- Verified all changes via static analysis and syntax checks.

+++ database.sql
@@ -1514,7 +1514,7 @@ CREATE TABLE `idf_positions` (
   `id` int NOT NULL AUTO_INCREMENT,
   `company_id` int NOT NULL,
   `idf_id` int NOT NULL,
-  `position_no` tinyint NOT NULL,
+  `position_no` smallint NOT NULL,
   `device_type` int NOT NULL,
   `device_name` varchar(140) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
   `equipment_id` varchar(9) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
